add cast
* hash.c (rb_objid_hash): need to cast down.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@56657 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ hash.c
@@ -271,7 +271,7 @@ key64_hash (uint64_t key, uint32_t seed)
 long
 rb_objid_hash(st_index_t index)
 {
-    return key64_hash(index, (uint32_t) prime2);
+    return (long)key64_hash(index, (uint32_t)prime2);
 }
 
 static st_index_t
